requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Pillow==10.1.0
ffmpeg-python==0.2.0
PyPDF2==3.0.1
Ghostscript==0.7
qpdf==1.3.0
pikepdf==8.7.1